refactor: split to accelerator, keyboard-event, local and global supported keys, move global and local shortcut services
