--- apps/app/src/lib/services/shortcuts/formatConverters.ts
@@ -1,17 +1,0 @@
-import type { SupportedKey } from './createLocalShortcutManager';
-
-/**
- * Convert a shortcut string to an array of keys
- * @example "ctrl+shift+a" → ["ctrl", "shift", "a"]
- */
-export function shortcutStringToArray(shortcut: string): SupportedKey[] {
-	return shortcut.split('+').map((key) => key.toLowerCase() as SupportedKey);
-}
-
-/**
- * Join an array of keys into a shortcut string
- * @example ["ctrl", "shift", "a"] → "ctrl+shift+a"
- */
-export function arrayToShortcutString(keys: SupportedKey[]): string {
-	return keys.map((key) => key.toLowerCase()).join('+');
-}
